fITness_backend/requirements.txt
Django==5.0
mysqlclient==2.2.0
djangorestframework